+++ packages/contracts-communication/script/config/ConfigureAppV1.s.sol
@@ -8,8 +8,6 @@ import {TypeCasts} from "../../contracts/libs/TypeCasts.sol";
 
 import {stdJson, SynapseScript} from "@synapsecns/solidity-devops/src/SynapseScript.sol";
 
-// solhint-disable code-complexity
-// solhint-disable custom-errors
 abstract contract ConfigureAppV1 is SynapseScript {
     using stdJson for string;
 
@@ -24,6 +22,7 @@ abstract contract ConfigureAppV1 is SynapseScript {
 
     function run(string memory environment) external virtual broadcastWithHooks {
         loadConfig(environment);
+        beforeAppConfigured();
         linkRemoteChains();
         syncTrustedModules();
         setAppConfig();
@@ -175,6 +174,9 @@ abstract contract ConfigureAppV1 is SynapseScript {
             printLog(string.concat("Added ", vm.toString(added), " clients, removed ", vm.toString(removed)));
         }
     }
+
+    // solhint-disable-next-line no-empty-blocks
+    function beforeAppConfigured() internal virtual {}
 
     // solhint-disable-next-line no-empty-blocks
     function afterAppConfigured() internal virtual {}