+++ packages/synapse-interface/utils/joinClassNames.ts
@@ -1,0 +1,1 @@
+export const joinClassNames = (a) => Object.values(a).join(' ')